Implementacion de response entities, mapper en PersonaController y PersonaService

+++ src/main/java/com/example/demo/services/PersonaService.java
@@ -1,15 +1,22 @@
 package com.example.demo.services;
 
+import ch.qos.logback.classic.Logger;
+import com.example.demo.dtos.PersonaRequest;
 import com.example.demo.dtos.PersonaResponse;
 import com.example.demo.dtos.PersonasResponse;
 import com.example.demo.entities.Persona;
 import com.example.demo.mappers.PersonaMapper;
 import com.example.demo.repositories.IPersonaRepository;
+import jakarta.transaction.Transactional;
+import org.apache.velocity.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.ResponseEntity;
+
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -17,41 +24,66 @@ public class PersonaService {
 
     @Autowired
     IPersonaRepository repository;
-
     @Autowired
     PersonaMapper mapper;
-
-    public List<Persona> getPersonas(){
-        return repository.findAll();
-        //return discoMapper.ListDiscosToDiscosResponse(discoRepository.findAll());
-    }
-
-    public Optional<Persona> getById(Long id){
-        return (Optional<Persona>) repository.findById(id);
-    }
-
-    public ResponseEntity savePersona(Persona persona){
-        repository.save(persona);
-        return ResponseEntity.ok("Registro guardado");
-    }
-/*
-    public Persona updateById(Persona request, Long id){
-        Persona disco = discoRepository.findById(id).get();
-        disco.setTitulo(request.getTitulo());
-        disco.setFechaPublicacion(request.getFechaPublicacion());
-
-        saveDisco(disco);
-        return disco;
-    }
-*/
-
-    public boolean deletePersonaById(Long id){
-        try{
-            repository.deleteById(id);
-            return true;
-        }catch (Exception e){
-            return false;
+    @Transactional
+    public ResponseEntity<PersonasResponse> getPersonas(){
+        try {
+            return ResponseEntity.ok().body(mapper.ListToResponse(repository.findAll()));
+        } catch (Exception e) {
+            return ResponseEntity.internalServerError().build();
         }
     }
+    @Transactional
+    public ResponseEntity<PersonaResponse> getById(Long id) {
+        try {
+            Persona persona = repository.findById(id).get();
+            return ResponseEntity.ok(mapper.ModelToResponse(persona));
+        } catch (NoSuchElementException e) {
+            return ResponseEntity.notFound().build();
+        }
+    }
+    @Transactional
+    public ResponseEntity savePersona(PersonaRequest persona) {
+        try {
+            repository.save(mapper.RequestToModel(persona));
+            return ResponseEntity.ok("Registro guardado");
+        } catch (DataIntegrityViolationException e) {
+            return ResponseEntity.badRequest().body("Error al guardar el registro: " + e.getMessage());
+        }
+    }
+    @Transactional
+    public ResponseEntity<PersonaResponse> updateById(PersonaRequest request, Long id) {
+        try {
+            Persona persona = repository.findById(id).orElseThrow(() ->
+                    new ResourceNotFoundException("Persona no encontrada con id: " + id));
+
+            persona.setDni(request.getDni());
+            persona.setNombre(request.getNombre());
+            persona.setFechaNacimiento(request.getFechaNacimiento());
+            repository.save(persona);
+            return ResponseEntity.ok(mapper.ModelToResponse(persona));
+        } catch (ResourceNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        } catch (Exception e) {
+            return ResponseEntity.internalServerError().build();
+        }
+    }
+
+
+    @Transactional
+    public ResponseEntity<String> deletePersonaById(Long id) {
+        try {
+            if (repository.existsById(id)) {
+                repository.deleteById(id);
+                return ResponseEntity.ok("Persona eliminada");
+            } else {
+                return ResponseEntity.notFound().build();
+            }
+        } catch (Exception e) {
+            return ResponseEntity.internalServerError().body("Error interno del servidor");
+        }
+    }
+
 
 }
